fix: Ensure consistent Delta time type as Uint64 across scheduler and processes

+++ src/Processes/InputProcess.hpp
@@ -3,7 +3,7 @@
 #include <entt/process/process.hpp>
 #include <SDL2/SDL.h>
 
-class InputProcess : public entt::process<InputProcess, float> {
+class InputProcess : public entt::process<InputProcess, Uint64> {
 public:
     InputProcess(entt::registry& registry)
         : registry_(registry) {}
